Fix - 'Expense - Report RHP re-renders every time clicking Create expense'
